changed code structure
