--- src/Coder/Coder.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import CoderContainer from './containers/CoderContainer'
-
-const Coder = () => {
-  return (
-    <CoderContainer />
-  )
-}
-
-export default Coder
