--- src/ReviewList/ReviewList.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import ReviewListContainer from './containers/ReviewListContainer'
-
-const ReviewList = () => {
-  return (
-    <ReviewListContainer />
-  )
-}
-
-export default ReviewList
