+++ src/containers/Login.js
@@ -1,9 +1,17 @@
+// React Redux
 import React from 'react'
+import { connect } from 'react-redux'
+
+// Material UI
+import { withStyles } from '@material-ui/core/styles';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import ListItem from '@material-ui/core/ListItem';
 
-class CoderComponent extends React.Component {
+// ActionCreators
+import * as CoderActions from '../actions/CoderActions'
+
+class Login extends React.Component {
 
     handleChangeAccessToken = () => event => {
         this.props.changeCoderAccessToken(event.target.value)
@@ -12,7 +20,7 @@ class CoderComponent extends React.Component {
     handleLoginButtonClicked = () => event => {
         this.props.fetchCoderInfo(this.props.coder_info.access_token)
     }
-    
+
     render() {
         return (
             <ListItem>
@@ -35,4 +43,19 @@ class CoderComponent extends React.Component {
     }
 }
 
-export default CoderComponent
+const mapStateToProps = (state) => {
+    return {
+        coder_info: state.coder_info
+    }
+}
+
+const mapDispatchToProps = (dispatch) => {
+    return {
+        changeCoderAccessToken: (text) => dispatch(CoderActions.changeCoderAccessToken(text)),
+        fetchCoderInfo: (access_token) => dispatch(CoderActions.fetchCoderInfo(access_token))
+    }
+}
+
+const styles = {};
+
+export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(Login))
