+++ src/index.js
@@ -2,10 +2,8 @@ import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import CssBaseline from '@material-ui/core/CssBaseline';
-
 import store from './store'
-import ReviewList from './ReviewList/ReviewList'
-
+import ReviewList from './containers/ReviewList'
 render (
   <Provider store={store}>
       <React.Fragment>
