+++ src/store.js
@@ -4,12 +4,14 @@ import { composeWithDevTools } from 'redux-devtools-extension'
 import logger from "redux-logger"
 import thunk from "redux-thunk"
 // reducers
-import ReviewListReducer from './ReviewList/ReviewListReducer'
-import CoderReducer from './Coder/CoderReducer'
+import CoderInfoReducer from './reducers/CoderInfoReducer'
+import InReviewProblemsReducer from './reducers/InReviewProblemsReducer'
+import ReviewQueueReducer from './reducers/ReviewQueueReducer'
 
 const reducers = combineReducers({
-    review_list: ReviewListReducer,
-    coder: CoderReducer
+    coder_info: CoderInfoReducer,
+    in_review_problems: InReviewProblemsReducer,
+    review_queue: ReviewQueueReducer
 })
 
 const middlewares = composeWithDevTools(
